+++ libp2p/pubsub/gossipsub.py
@@ -32,7 +32,7 @@ from libp2p.peer.id import (
 from libp2p.pubsub import (
     floodsub,
 )
-from libp2p.tools.async_service import (
+from libp2p.tools.anyio_service import (
     Service,
 )
 from libp2p.utils import (
