+++ libp2p/tools/anyio_service/exceptions.py
@@ -1,0 +1,16 @@
+class ServiceException(Exception):
+    """
+    Base class for Service exceptions
+    """
+
+
+class LifecycleError(ServiceException):
+    """
+    Raised when an action would violate the service lifecycle rules.
+    """
+
+
+class DaemonTaskExit(ServiceException):
+    """
+    Raised when a daemon task exits unexpectedly.
+    """
