+++ libp2p/tools/anyio_service/stats.py
@@ -1,0 +1,12 @@
+from dataclasses import dataclass
+
+
+@dataclass
+class TaskStats:
+    total_count: int
+    finished_count: int
+
+
+@dataclass
+class Stats:
+    tasks: TaskStats
