--- libp2p/tools/async_service/__init__.py
@@ -1,15 +1,0 @@
-from .abc import (
-    ServiceAPI,
-)
-from .base import (
-    Service,
-    as_service,
-)
-from .exceptions import (
-    DaemonTaskExit,
-    LifecycleError,
-)
-from .trio_service import (
-    TrioManager,
-    background_trio_service,
-)
